refactor(web): split DTO types out of lib/api.ts into lib/apiTypes.ts (#960 M1)
Pure motion: move all 153 exported DTO interface/type declarations, the
unexported `PlanSource`, and the two contract constants
(DEFAULT_AUTOPILOT_LABEL, RATE_LIMIT_SOURCES) out of web/src/lib/api.ts into
a new web/src/lib/apiTypes.ts. api.ts re-exports them via `export type *`
plus a value re-export so all 207 lib/api barrel importers stay byte-identical.
api.ts keeps MOCK_MODE, ProjectSyncVisibility, the two leaf re-exports, the
socket/client helpers, request, realApi and api. Comment repoints in
knip.jsonc, AdminBranding.tsx, OccurrenceFileIssue.tsx.

Part of PRD #960 (epic #915, finding W4).
